+++ contracts/babylon/src/msg/mod.rs
@@ -1,4 +1,4 @@
+pub mod consumer_header;
 pub mod contract;
-pub mod cz_header;
 pub mod epoch;
 pub mod ibc;
