+++ contracts/babylon/src/utils/mod.rs
@@ -5,4 +5,4 @@ mod cosmos_store;
 pub mod babylon_epoch_chain;
 pub mod btc_light_client_executor;
 pub mod btc_light_client_querier;
-pub mod cz_header_chain;
+pub mod consumer_header_chain;
