+++ contracts/btc-finality/src/state/public_randomness.rs
@@ -5,7 +5,7 @@ use cw_storage_plus::{Bound, Map};
 use crate::error::ContractError;
 use crate::state::config::CONFIG;
 use babylon_apis::finality_api::PubRandCommit;
-use babylon_contract::msg::cz_header::CzHeightResponse;
+use babylon_contract::msg::consumer_header::CzHeightResponse;
 
 /// Map of public randomness commitments by fp and block height
 pub const PUB_RAND_COMMITS: Map<(&str, u64), PubRandCommit> = Map::new("fp_pub_rand_commit");
